move OQIndexExpression to classical_types
